+++ pages/login/login.js
@@ -1,6 +1,5 @@
 import LoginController from './loginController.js';
 import LoginView from './loginView.js';
-import LoginService from './loginService.js';
 import user from '../domain/user.js';
 
 document.addEventListener('DOMContentLoaded', () => {
@@ -10,6 +9,5 @@ document.addEventListener('DOMContentLoaded', () => {
     return;
   }
   const loginView = new LoginView();
-  const loginService = new LoginService(loginView, user);
-  new LoginController(loginView, user, loginService);
+  new LoginController(loginView, user);
 });
